+++ examples/minimal/provider.tf
@@ -1,10 +1,13 @@
 terraform {
-  required_version = ">= 1.11.1"
-
   required_providers {
     github = {
       source  = "integrations/github"
       version = "~> 6.13.0"
     }
   }
+}
+
+provider "github" {
+  owner = var.github_owner
+  token = var.github_token
 }