+++ CHANGELOG.md
@@ -8,3 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## v1.0.0 - 2026-08-03
 
 * feature: Initialize GitHub Organizations module (#1)
+* feature: Add support for managing organization settings and GitHub Actions permissions
+* feature: Add support for managing teams, members, and team repository permissions
+* feature: Add support for managing repositories, default branches, labels, and branch protection
+* feature: Add support for managing organization secrets, Dependabot secrets, and webhooks